Refactored rendering code
